requirements.txt
django 
djangorestframework
psycopg2-binary
Pillow
python-decouple
Pillow
djangorestframework_simplejwt
dj-config-url